some commit changes

+++ Assignment 6 (struct datatype)/Examresults.cpp
@@ -3,9 +3,9 @@
 // Declare student data
 struct Student{
 	std::string Nim;
-	int Uts;
-	int Uas;
-	int avg;
+	float Uts;
+	float Uas;
+	float avg;
 };
 
 int main () {
